pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "psdl-lang"
version = "0.2.1"
description = "Patient Scenario Definition Language - Reference Implementation"
readme = "README_PYPI.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Ziyuan Guan", email = "ziyuan.guan@ufl.edu"}
]
maintainers = [
    {name = "Ziyuan Guan", email = "ziyuan.guan@ufl.edu"}
]
keywords = [
    "healthcare",
    "clinical",
    "patient",
    "scenario",
    "DSL",
    "OMOP",
    "FHIR",
    "EHR",
    "temporal",
    "medical"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Full runtime with all adapters
full = [
    "requests>=2.28",
    "sqlalchemy>=2.0",
]
# OMOP CDM adapter
omop = [
    "sqlalchemy>=2.0",
]
# FHIR R4 adapter
fhir = [
    "requests>=2.28",
]
# Streaming execution (Apache Flink)
streaming = [
    "apache-flink>=1.17",
]
# Development and testing
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "jsonschema>=4.0",
    "lark>=1.1",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
    "requests>=2.28",  # Required for FHIR backend tests
]
# Documentation
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.3",
    "myst-parser>=2.0",
]
# Jupyter notebook support
notebook = [
    "jupyter>=1.0",
    "ipywidgets>=8.0",
    "pandas>=2.0",
    "matplotlib>=3.7",
]
# All optional dependencies
all = [
    "psdl-lang[full,streaming,dev,docs,notebook]",
]

[project.urls]
Homepage = "https://github.com/Chesterguan/PSDL"
Documentation = "https://github.com/Chesterguan/PSDL#readme"
Repository = "https://github.com/Chesterguan/PSDL"
Issues = "https://github.com/Chesterguan/PSDL/issues"
Changelog = "https://github.com/Chesterguan/PSDL/blob/main/CHANGELOG.md"

[project.scripts]
psdl = "psdl.cli:main"

[project.entry-points."psdl.adapters"]
omop = "psdl.adapters.omop:OMOPAdapter"
fhir = "psdl.adapters.fhir:FHIRAdapter"

[tool.setuptools.packages.find]
where = ["src"]
include = ["psdl*"]

[tool.setuptools.package-data]
psdl = ["py.typed", "*.yaml", "*.json", "*.lark", "*.ebnf"]
"psdl.examples" = ["*.yaml"]

# ============================================================
# Tool Configuration
# ============================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (may require external services)",
    "slow: marks tests as slow running",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/psdl"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | venv
    | build
    | dist
    | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["psdl"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
